Optimised Delivery Method. Read Desc
As it turns out, setting the object active and inactive would cause spikes. Now they are always active. This is not fully tested, however any bugs that appear should be easy to solve. This method is impact 0 and should be kept at all costs. Other things will need to be added and sorted to accommodate this, but I believe it shouldn't be too bad. I also need to clean up a lot lol